refactor: `ctf_gpt.py` 분리
`ctf_gpt.py` 파일의 크기가 켜져 가독성 및 유지보수가 어려워짐. 이에 따라 파일을 기능별 책임에 따라 여러 개의 작은 모듈로 분리함.
이번 리팩토링은 기존 클래스 구조나 로직 변경을 최소화하면서 코드 구조를 개선하는 데 초점을 맞춤.

- **`utils/` 경로에 신규 핸들러 모듈 생성:**
  - `cg_llm_handlers.py`: LLM 에이전트와의 모든 상호작용 로직을 담당.
  - `cg_ui_handlers.py`: 사용자 입력 프롬프트 및 모든 UI 관련 로직을 담당.
  - `cg_session_handlers.py`: 세션 로드, 저장, 대화 로깅 등 세션 생명주기를 관리.
  - `cg_auto_handlers.py`: `auto` 명령어를 위한 자동화된 도구 실행 로직을 담당.
  - `cg_constants.py`: 공유 상수 및 간단한 유틸리티 함수를 중앙에서 관리.

- **`ctf_gpt.py` 수정:**
  - `CTFGPT` 클래스가 신규 모듈의 핸들러 함수들을 자신의 메서드로 동적 바인딩하도록 수정.
  - 핵심 `__init__`과 `main` 루프는 유지하되, 내부 로직을 각 핸들러에 위임하여 복잡도를 낮춤.
