+++ utils/cg_auto_handlers.py
@@ -1,0 +1,227 @@
+import json
+import asyncio
+import re
+from prompt_toolkit.shortcuts import confirm
+from prompt_toolkit.formatted_text import HTML
+from utils.prompt_select import prompt_select
+
+async def run_single_auto_step(self, interactive: bool):
+    """
+    Executes a single automated step (tool execution) based on the LLM's suggestion.
+    The behavior differs based on the 'interactive' flag.
+    :param interactive: If True, the user selects the tool and confirms arguments.
+                        If False, the LLM selects the server, tool, and arguments.
+    """
+    self.console.print("\n--- Preparing for automated step ---", style="bold magenta")
+    
+    if not self.step_reasoning:
+        self.console.print("No task has been suggested yet. Please run 'next' or 'todo' first.", style="bold red")
+        return
+    
+    task_to_execute = self.step_reasoning
+
+    # 1. Fetch all available tools from all servers first.
+    with self.console.status("[bold green]Fetching tools from all available MCP servers..."):
+        all_tools_by_server = {}
+        server_names = self.mcp_manager.get_available_servers()
+        for server_name in server_names:
+            client = self.mcp_manager.create_client(server_name)
+            if not client: continue
+            try:
+                async with client:
+                    # In fastmcp, is_connected() is a property, not an async method
+                    if client.is_connected:
+                        tools = await client.list_tools()
+                        all_tools_by_server[server_name] = [
+                            {"name": tool.name, "description": tool.description, "inputSchema": tool.inputSchema} for tool in tools
+                        ]
+            except Exception as e:
+                self.console.print(f"Warning: Could not fetch tools from '{server_name}': {e}", style="yellow")
+    
+    if not all_tools_by_server:
+        self.console.print("Could not fetch any tools from any MCP server. Aborting.", style="bold red")
+        return
+
+    action = {}
+    # 2. Branch logic based on interactive mode
+    if interactive:
+        # --- Interactive Mode ---
+        # 2a. User selects the tool
+        self.console.print("Please select a tool to use for the task:", style="bold green")
+        tool_options = []
+        for server, tools in all_tools_by_server.items():
+            for tool in tools:
+                tool_id = f"{server}.{tool['name']}"
+                display_text = HTML(f'<style fg="cyan">{server}</style>.<style fg="yellow">{tool["name"]}</style>: {tool["description"]}')
+                tool_options.append((tool_id, display_text))
+
+        selected_tool_id = await asyncio.to_thread(prompt_select, title="Available Tools:", values=tool_options)
+        
+        if not selected_tool_id:
+            self.console.print("No tool selected. Aborting auto step.", style="yellow")
+            return
+        
+        selected_server_name, selected_tool_name = selected_tool_id.split('.', 1)
+        selected_tool_info = next(t for t in all_tools_by_server[selected_server_name] if t['name'] == selected_tool_name)
+
+        # 2b. LLM generates arguments for the selected tool
+        prompt_for_args = f"""
+        You are an expert at generating command arguments.
+        The user wants to perform the following task:
+        ---
+        {task_to_execute}
+        ---
+        The user has chosen to use the '{selected_tool_name}' tool from the '{selected_server_name}' server.
+        Here is the tool's information:
+        ```json
+        {json.dumps(selected_tool_info, indent=2)}
+        ```
+        Based on the task and the tool's input schema, generate the appropriate JSON object for the 'arguments' key.
+        Your output MUST be only a single JSON object containing the arguments.
+
+        Example output for nmap_scan:
+        {{"targets": "localhost", "options": "-p-"}}
+        """
+        if False:
+            # --- DEBUG: Print prompt for LLM ---
+            self.console.print("\n--- DEBUG: Prompt to LLM ---", style="yellow")
+            self.console.print(prompt_for_args)
+            self.console.print("--------------------------\n", style="yellow")
+
+        with self.console.status(f"[bold green]LLM is generating arguments for {selected_tool_name}..."):
+            args_json_str = await asyncio.to_thread(self.reasoningAgent.send_message, prompt_for_args, self.test_reasoning_session_id)
+        if False:
+            # --- DEBUG: Print raw LLM response ---
+            self.console.print("\n--- DEBUG: Raw LLM Response ---", style="yellow")
+            self.console.print(args_json_str)
+            self.console.print("-----------------------------\n", style="yellow")
+        
+        try:
+            json_match = re.search(r'\{.*\}', args_json_str, re.DOTALL)
+            if not json_match: raise ValueError("No JSON object found in the response.")
+            arguments = json.loads(json_match.group())
+            action = {
+                "server_name": selected_server_name,
+                "tool_name": selected_tool_name,
+                "arguments": arguments
+            }
+        except (json.JSONDecodeError, AttributeError, ValueError) as e:
+            self.console.print(f"Failed to parse arguments from LLM: {e}", style="bold red")
+            self.console.print("LLM Response:", args_json_str, style="red")
+            return
+
+    else:
+        # --- Fully Automated Mode ---
+        # 2c. LLM selects server, tool, and arguments in one go
+        all_tools_json_str = json.dumps(all_tools_by_server, indent=2)
+        prompt_for_full_action = f"""
+        You are an expert CTF assistant responsible for selecting the best tool for a given task from multiple available tool servers.
+
+        Recommended Task:
+        ---
+        {task_to_execute}
+        ---
+
+        Available Tool Servers and their Tools:
+        ```json
+        {all_tools_json_str}
+        ```
+
+        Based on the "Recommended Task", analyze the tools available on each server. Select the single most appropriate server and tool to execute the task.
+        Your output MUST be a single JSON object with "server_name", "tool_name", and "arguments" keys.
+
+        IMPORTANT: If no tool on any server is suitable, return a JSON object with "server_name" set to "none".
+
+        Example for a port scan task:
+        {{"server_name": "ctf-tools", "tool_name": "nmap_scan", "arguments": {{"targets": "localhost", "options": "-p-"}}}}
+
+        Example for when no tool is suitable:
+        {{"server_name": "none", "tool_name": "none", "arguments": {{"reason": "The task requires a specific vulnerability scanner that is not available."}}}}
+        """
+        if False:
+            # --- DEBUG: Print prompt for LLM ---
+            self.console.print("\n--- DEBUG: Prompt to LLM ---", style="yellow")
+            self.console.print(prompt_for_full_action)
+            self.console.print("--------------------------\n", style="yellow")
+
+        with self.console.status("[bold green]LLM is selecting the best server, tool, and arguments..."):
+            action_json_str = await asyncio.to_thread(self.reasoningAgent.send_message, prompt_for_full_action, self.test_reasoning_session_id)
+        if False:
+            # --- DEBUG: Print raw LLM response ---
+            self.console.print("\n--- DEBUG: Raw LLM Response ---", style="yellow")
+            self.console.print(action_json_str)
+            self.console.print("-----------------------------\n", style="yellow")
+
+        try:
+            json_match = re.search(r'\{.*\}', action_json_str, re.DOTALL)
+            if not json_match: raise ValueError("No JSON object found in the response.")
+            action = json.loads(json_match.group())
+        except (json.JSONDecodeError, AttributeError, ValueError) as e:
+            self.console.print(f"Failed to parse action from LLM: {e}", style="bold red")
+            self.console.print("LLM Response:", action_json_str, style="red")
+            return
+
+    # 3. Execution (common for both modes)
+    server_name = action.get("server_name")
+    tool_name = action.get("tool_name")
+    arguments = action.get("arguments")
+
+    if not server_name or server_name == 'none':
+        self.console.print("LLM determined no suitable tool is available or chose not to act.", style="yellow")
+        return
+
+    self.console.print(f"Suggested Action: Run tool '{tool_name}' on server '{server_name}' with arguments: {arguments}", style="cyan")
+
+    # Final confirmation step (always happens in interactive mode, skipped in full-auto)
+    execute = True
+    if interactive:
+        execute = await asyncio.to_thread(confirm, f"Proceed with execution?")
+    
+    if not execute:
+        self.console.print("Execution cancelled by user.", style="yellow")
+        return
+
+    # 4. Connect to the chosen server and run the tool
+    client = self.mcp_manager.create_client(server_name)
+    if not client:
+        self.console.print(f"Failed to create client for the selected server '{server_name}'.", style="bold red")
+        return
+    
+    tool_output_text = "No output received from tool."
+    async with client:
+        if not client.is_connected:
+            self.console.print(f"MCP Client could not connect to '{server_name}'. Check server script and permissions.", style="bold red")
+            return
+        
+        with self.console.status(f"[bold green]Executing '{tool_name}' on '{server_name}'..."):
+            try:
+                tool_results_list = await client.call_tool(tool_name, arguments)
+                if not tool_results_list:
+                    tool_output_text = f"Tool '{tool_name}' executed but returned no result."
+                    self.console.print(tool_output_text, style="yellow")
+                else:
+                    tool_output_text = "\n".join([result.text for result in tool_results_list])
+            except Exception as e:
+                tool_output_text = f"Error during tool execution: {e}"
+                self.console.print(tool_output_text, style="bold red")
+
+    # Log the results
+    mcp_log_entry = {
+        "server": server_name, 
+        "tool": tool_name,
+        "arguments": arguments,
+        "result": tool_output_text
+    }
+    self.log_conversation("mcp_tool_log", json.dumps(mcp_log_entry, indent=2))
+    
+    self.console.print(f"--- Tool '{tool_name}' Output ---", style="green")
+    self.console.print(tool_output_text)
+    self.console.print("--- End of Output ---", style="green")
+    
+    # 5. Analyze results and update CTT
+    with self.console.status("[bold green]Analyzing results and updating CTT..."):
+        self.step_reasoning = await asyncio.to_thread(self.reasoning_handler, tool_output_text)
+        self.console.print("\nCTFGPT's analysis and next recommended steps:", style="bold green")
+        self.console.print(self.step_reasoning)
+
+    self.console.print("CTT updated.", style="green")
