+++ utils/cg_constants.py
@@ -1,0 +1,30 @@
+from prompt_toolkit.formatted_text import HTML
+
+# Constants for ctfGPT
+postfix_options = {
+    "tool": "The input content is from a security testing tool. You need to list down all the points that are interesting to you; you should summarize it as if you are reporting to a senior penetration tester for further guidance.\n",
+    "user-comments": "The input content is from user comments.\n",
+    "web": "The input content is from web pages. You need to summarize the readable-contents, and list down all the points that can be interesting for penetration testing.\n",
+    "default": "The user did not specify the input source. You need to summarize based on the contents.\n",
+}
+
+options_desc = {
+    "tool": " Paste the output of the security test tool used",
+    "user-comments": "",
+    "web": " Paste the relevant content of a web page",
+    "default": " Write whatever you want, the tool will handle it",
+}
+
+def prompt_continuation(width, line_number, wrap_count):
+    """
+    The continuation: display line numbers and '->' before soft wraps.
+    Notice that we can return any kind of formatted text from here.
+    The prompt continuation doesn't have to be the same width as the prompt
+    which is displayed before the first line, but in this example we choose to
+    align them. The `width` input that we receive here represents the width of
+    the prompt.
+    """
+    if wrap_count > 0:
+        return " " * (width - 3) + "-> "
+    text = ("- %i - " % (line_number + 1)).rjust(width)
+    return HTML("<strong>%s</strong>") % text
