+++ utils/cg_llm_handlers.py
@@ -1,0 +1,99 @@
+import textwrap
+from prompts.prompt_class import CTFGPTPrompt 
+
+# Placeholder for actual logger, should be passed or initialized in CTFGPT class
+# from loguru import logger 
+
+def _feed_init_prompts(self):
+    # 1. User firstly provide basic information of the task
+    init_description = self.prompt_ask(
+        "Please describe the CTF challenge in one line, including the target IP, task type, etc.\n> ",
+        multiline=False,
+    )
+    self.log_conversation("user", init_description)
+    self.task_log["task description"] = init_description
+    # 2. Provide the information to the reasoning session for the task initialization.
+    # Note that this information is not parsed by the three-step process in reasoning.
+    # It is directly used to initialize the task.
+    prefixed_init_description = self.prompts.task_description + init_description
+    with self.console.status(
+        "[bold green] Constructing Initial CTF Task Tree..."
+    ) as status:
+        _reasoning_response = self.reasoningAgent.send_message(
+            prefixed_init_description, self.test_reasoning_session_id
+        )
+    # 3. Pass to generation session for more details.
+    # Note that the generation session is not used for the task initialization.
+    with self.console.status("[bold green] Generating Initial Task") as status:
+        _generation_response = self.generationAgent.send_message(
+            self.prompts.todo_to_command + _reasoning_response,
+            self.test_generation_session_id,
+        )
+    
+    self.step_reasoning = _reasoning_response 
+
+    # Display the initial generation result
+    response = _reasoning_response + "\n" + _generation_response
+    self.console.print("CTFGPT output: ", style="bold green")
+    self.console.print(response)
+    self.log_conversation("ctfGPT", "CTFGPT output:" + response)
+
+def reasoning_handler(self, text) -> str:
+    # summarize the contents if necessary.
+    if len(text) > self.parsing_char_window:
+        text = self.input_parsing_handler(text)
+    """
+    # pass the information to reasoning_handler and obtain the results
+    response = self.reasoningAgent.send_message(
+        self.prompts.process_results + text, self.test_reasoning_session_id
+    )
+    # log the conversation
+    """
+    # CTFGPT Reasoning Logic
+    ## 1. Given the information, update the CTT
+    _updated_ctt_response = self.reasoningAgent.send_message(
+        self.prompts.process_results + text, self.test_reasoning_session_id
+    )
+    ## 2. Validate if the CTT is correct
+    # TODO
+    ## 3. If the CTT is correct, select all the to-dos
+    _task_selection_response = self.reasoningAgent.send_message(
+        self.prompts.process_results_task_selection, self.test_reasoning_session_id
+    )
+    # get the complete output:
+    response = _updated_ctt_response + _task_selection_response
+
+    self.log_conversation("reasoning", response)
+    return response
+
+def input_parsing_handler(self, text, source=None) -> str:
+    prefix = "Please summarize the following input. "
+    # do some engineering trick here. Add postfix to the input to make it more understandable by LLMs.
+    if source is not None and source in self.postfix_options.keys():
+        prefix += self.postfix_options[source]
+    # The default token-size limit is 4096 (web UI even shorter). 1 token ~= 4 chars in English
+    # Use textwrap to split inputs. Limit to 2000 token (8000 chars) for each input
+    # (1) replace all the newlines with spaces
+    text = text.replace("\r", " ").replace("\n", " ")
+    # (2) wrap the text
+    wrapped_text = textwrap.fill(text, 8000)
+    wrapped_inputs = wrapped_text.split("\n")
+    # (3) send the inputs to chatGPT input_parsing_session and obtain the results
+    summarized_content = ""
+    for wrapped_input in wrapped_inputs:
+        word_limit = f"Please ensure that the input is less than {8000 / len(wrapped_inputs)} words.\n"
+        summarized_content += self.parsingAgent.send_message(
+            prefix + word_limit + wrapped_input, self.input_parsing_session_id
+        )
+    # log the conversation
+    self.log_conversation("input_parsing", summarized_content)
+    return summarized_content
+
+def test_generation_handler(self, text):
+    # send the contents to chatGPT test_generation_session and obtain the results
+    response = self.generationAgent.send_message(
+        text, self.test_generation_session_id
+    )
+    # log the conversation
+    self.log_conversation("generation", response)
+    return response
