+++ utils/cg_session_handlers.py
@@ -1,0 +1,115 @@
+import json
+import os
+import time
+from prompt_toolkit.shortcuts import confirm, prompt
+
+from utils.prompt_select import prompt_ask
+
+def log_conversation(self, source, text):
+    """
+    append the conversation into the history
+
+    Parameters:
+    ----------
+    source: str
+        the source of the conversation
+    text: str
+        the content of the conversation
+    """
+    # append the conversation into the history
+    timestamp = time.time()
+    if source not in self.history.keys():
+        # an exception
+        source = "exception"
+    self.history[source].append((timestamp, text))
+
+def save_session(self):
+    """
+    Save the current session for next round of usage.
+    The test information is saved in the directory `./test_history`
+    """
+    self.console.print(
+        "Before you quit, you may want to save the current session.",
+        style="bold green",
+    )
+    # 1. Require a save name from the user. If not, use the current time as the save name.
+    save_name = prompt_ask(
+        "Please enter the name of the current session. (Default with current timestamp)\n> ",
+        multiline=False,
+    )
+    if save_name == "":
+        save_name = str(time.time())
+    
+    # Ensure the save directory exists
+    os.makedirs(self.save_dir, exist_ok=True)
+    
+    save_file_path = os.path.join(self.save_dir, save_name)
+
+    # 2. Save the current session
+    with open(save_file_path, "w") as f:
+        # store the three ids and task_log
+        session_ids = {
+            "reasoning": self.test_reasoning_session_id,
+            "test_generation": self.test_generation_session_id,
+            "parsing": self.input_parsing_session_id,
+            "task_log": self.task_log,
+        }
+        json.dump(session_ids, f)
+    self.console.print(
+        f"The current session is saved as {save_name} in '{self.save_dir}'", style="bold green"
+    )
+    return
+
+def _preload_session(self) -> dict:
+    """
+    Preload the session from the save directory.
+
+    Returns:
+        dict: the session ids for the three sessions.
+        None if no previous session is found.
+    """
+    if continue_from_previous := confirm(
+        "Do you want to continue from previous session?"
+    ):
+        # load the filenames from the save directory
+        if not os.path.exists(self.save_dir) or not os.path.isdir(self.save_dir):
+            self.console.print(f"Save directory '{self.save_dir}' not found. Starting a new session.", style="bold yellow")
+            return None
+            
+        filenames = os.listdir(self.save_dir)
+        if len(filenames) == 0:
+            print("No previous session found. Please start a new session.")
+            return None
+        else:  # print all the files
+            print("Please select the previous session by its index (integer):")
+            for i, filename in enumerate(filenames):
+                print(f"{str(i)}. {filename}")
+            # ask for the user input
+            try:
+                selected_index = input("Please key in your option (integer): ")
+                previous_testing_name = filenames[int(selected_index)]
+                print(f"You selected: {previous_testing_name}")
+            except (ValueError, IndexError) as e:
+                print("You input an invalid option. Will start a new session.")
+                return None
+
+    elif continue_from_previous is False:
+        return None
+    else:
+        print("You input an invalid option. Will start a new session.")
+        return None
+        
+    # 2. load the previous session information
+    if previous_testing_name is not None:
+        # try to load the file content with json
+        try:
+            with open(os.path.join(self.save_dir, previous_testing_name), "r") as f:
+                return json.load(f)
+        except Exception as e:
+            print(
+                "Error when loading the previous session. The file name is not correct"
+            )
+            print(e)
+            previous_testing_name = None
+            return None
+    return None # Fallback in case previous_testing_name is None after all checks
