+++ utils/cg_ui_handlers.py
@@ -1,0 +1,288 @@
+from prompt_toolkit.formatted_text import HTML
+from prompt_toolkit.shortcuts import confirm
+
+from utils.prompt_select import prompt_ask, prompt_select
+from utils.task_handler import local_task_entry, localTaskCompleter, main_task_entry, mainTaskCompleter
+from .cg_constants import postfix_options, options_desc 
+import asyncio
+
+def _select_mcp_server(self) -> str:
+    """사용 가능한 MCP 서버 목록을 표시하고 사용자에게 선택을 요청합니다."""
+    available_servers = self.mcp_manager.get_available_servers()
+    
+    if not available_servers:
+        self.console.print("No available MCP servers found in the configuration file.", style="bold red")
+        return None
+
+    if len(available_servers) == 1:
+        server = available_servers[0]
+        self.console.print(f"Automatically selecting the only available MCP server: '{server}'", style="bold yellow")
+        return server
+            
+    server_options = [(str(i), server) for i, server in enumerate(available_servers)]
+    selected_idx_str = prompt_select(
+        title="Please select an MCP server to use:",
+        values=server_options
+    )
+    
+    if selected_idx_str is None:
+        return None
+            
+    return available_servers[int(selected_idx_str)]
+
+def local_input_handler(self) -> str:
+    """
+    Request for user's input to handle the local task
+    """
+    local_task_response = ""
+    self.chat_count += 1
+    local_request_option = local_task_entry()
+    self.log_conversation("user", local_request_option)
+
+    if local_request_option == "help":
+        print(localTaskCompleter().task_details)
+
+    elif local_request_option == "discuss":
+        ## (1) Request for user multi-line input
+        self.console.print(
+            "Please share your findings and questions with CTFGPT."
+        )
+        self.log_conversation(
+            "ctfGPT",
+            "Please share your findings and questions with CTFGPT. (End with <shift + right-arrow>)",
+        )
+        user_input = prompt_ask("Your input: ", multiline=True)
+        self.log_conversation("user", user_input)
+        ## (2) pass the information to the reasoning session.
+        with self.console.status("[bold green] CTFGPT Thinking...") as status:
+            local_task_response = self.test_generation_handler(
+                self.prompts.local_task_prefix + user_input
+            )
+        ## (3) print the results
+        self.console.print("CTFGPT:\n", style="bold green")
+        self.console.print(local_task_response + "\n", style="yellow")
+        self.log_conversation("ctfGPT", local_task_response)
+
+    elif local_request_option == "brainstorm":
+        ## (1) Request for user multi-line input
+        self.console.print(
+            "Please share your concerns and questions with CTFGPT."
+        )
+        self.log_conversation(
+            "ctfGPT",
+            "Please share your concerns and questions with CTFGPT. End with <shift + right-arrow>)",
+        )
+        user_input = prompt_ask("Your input: ", multiline=True)
+        self.log_conversation("user", user_input)
+        ## (2) pass the information to the reasoning session.
+        with self.console.status("[bold green] CTFGPT Thinking...") as status:
+            local_task_response = self.test_generation_handler(
+                self.prompts.local_task_brainstorm + user_input
+            )
+        ## (3) print the results
+        self.console.print("CTFGPT:\n", style="bold green")
+        self.console.print(local_task_response + "\n", style="yellow")
+        self.log_conversation("ctfGPT", local_task_response)
+
+    elif local_request_option == "continue":
+        self.console.print("Exit the local task and continue the CTF.")
+        self.log_conversation(
+            "ctfGPT", "Exit the local task and continue the CTF."
+        )
+        local_task_response = "continue"
+
+    return local_task_response
+
+def input_handler(self) -> str:
+    """
+    Request for user's input to:
+        (1) input test results,
+        (2) ask for todos,
+        (3) input other information (discuss),
+        (4) end.
+    The design details are based on CTFGPT_design.md
+
+    Return
+    -----
+    response: str
+        The response from the chatGPT model.
+    """
+    self.chat_count += 1
+
+    request_option = main_task_entry()
+    self.log_conversation("user", request_option)
+    response = ""
+
+    if request_option == "help":
+        print(mainTaskCompleter().task_details)
+        response = "Displayed help."
+
+    if request_option == "next":
+        ## (1) pass the information to input_parsing session.
+        ## Give an option list for user to choose from
+        options = list(self.postfix_options.keys())
+        opt_desc = list(self.options_desc.values())
+
+        value_list = [
+            (
+                str(i),
+                HTML(
+                    f'<style fg="cyan">{options[i]}</style><style fg="LightSeaGreen">{opt_desc[i]}</style>'
+                ),
+            )
+            for i in range(len(options))
+        ]
+        source_idx_str = prompt_select(
+            title="Please choose the source of the information.", values=value_list
+        )
+        if source_idx_str is None:
+            return "Operation cancelled."
+        source = options[int(source_idx_str)]
+        
+        self.console.print(
+            "Your input: (End with <shift + right-arrow>)", style="bold green"
+        )
+        user_input = prompt_ask("> ", multiline=True)
+        self.log_conversation(
+            "user", f"Source: {source}" + "\n" + user_input
+        )
+        with self.console.status("[bold green] CTFGPT Thinking...") as status:
+            parsed_input = self.input_parsing_handler(
+                user_input, source=source
+            )
+            ## (2) pass the summarized information to the reasoning session.
+            reasoning_response = self.reasoning_handler(parsed_input)
+            self.step_reasoning = reasoning_response
+
+        ## (3) print the results
+        self.console.print(
+            "Based on the analysis, the following tasks are recommended:",
+            style="bold green",
+        )
+        self.console.print(reasoning_response + "\n")
+        self.log_conversation(
+            "ctfGPT",
+            "Based on the analysis, the following tasks are recommended:"
+            + reasoning_response,
+        )
+        response = reasoning_response
+
+    elif request_option == "more":
+        self.log_conversation("user", "more")
+        ## (1) check if reasoning session is initialized
+        if not self.step_reasoning:
+            response = "You have not initialized the task yet. Please perform the basic testing following `next` option."
+            self.console.print(response, style="bold red")
+            self.log_conversation("ctfGPT", response)
+            return response
+        ## (2) start local task generation.
+        ### (2.1) ask the reasoning session to analyze the current situation, and explain the task
+        self.console.print(
+            "CTFGPT will generate more step details, and enter the sub-task generation mode. (Pressing Enter to continue)",
+            style="bold green",
+        )
+        self.log_conversation(
+            "ctfGPT",
+            "CTFGPT will generate more step details, and enter the sub-task generation mode.",
+        )
+        input()
+
+        ### (2.2) pass the sub-tasks to the test generation session
+        with self.console.status("[bold green] CTFGPT Thinking...") as status:
+            generation_response = self.test_generation_handler(
+                self.step_reasoning
+            )
+            # Prime the generation agent for the upcoming local task session.
+            self.test_generation_handler(self.prompts.local_task_init)
+
+        self.console.print(
+            "Below are the further details.",
+            style="bold green",
+        )
+        self.console.print(generation_response + "\n")
+        response = generation_response
+        self.log_conversation("ctfGPT", response)
+
+        ### (2.3) local task handler
+        while True:
+            local_task_response = self.local_input_handler()
+            if local_task_response == "continue":
+                # break the local task handler
+                break
+
+    elif request_option == "todo":
+        ## log that user is asking for todo list
+        self.log_conversation("user", "todo")
+        ## (1) ask the reasoning session to analyze the current situation, and list the top sub-tasks
+        with self.console.status("[bold green] CTFGPT Thinking...") as status:
+            reasoning_response = self.reasoning_handler(self.prompts.ask_todo)
+            self.step_reasoning = reasoning_response 
+            ## (2) pass the sub-tasks to the test_generation session.
+            message = self.prompts.todo_to_command + "\n" + reasoning_response
+            generation_response = self.test_generation_handler(message)
+        ## (3) print the results
+        self.console.print(
+            "Based on the analysis, the following tasks are recommended:",
+            style="bold green",
+        )
+        self.console.print(reasoning_response + "\n")
+        self.console.print(
+            "You can follow the instructions below to complete the tasks.",
+            style="bold green",
+        )
+        self.console.print(generation_response + "\n")
+        response = reasoning_response
+        log_msg = (
+            "Based on the analysis, the following tasks are recommended:" + response + "\n" +
+            "You can follow the instructions below to complete the tasks." + generation_response
+        )
+        self.log_conversation("ctfGPT", log_msg)
+
+    elif request_option == "discuss":
+        ## (1) Request for user multi-line input
+        self.console.print(
+            "Please share your thoughts/questions with CTFGPT. (End with <shift + right-arrow>) "
+        )
+        self.log_conversation(
+            "ctfGPT", "Please share your thoughts/questions with CTFGPT."
+        )
+        user_input = prompt_ask("Your input: ", multiline=True)
+        self.log_conversation("user", user_input)
+        ## (2) pass the information to the reasoning session.
+        with self.console.status("[bold green] CTFGPT Thinking...") as status:
+            response = self.reasoning_handler(self.prompts.discussion + user_input)
+        ## (3) print the results
+        self.console.print("CTFGPT:\n", style="bold green")
+        self.console.print(response + "\n", style="yellow")
+        self.log_conversation("ctfGPT", response)
+
+    elif request_option == "auto":
+        self.log_conversation("user", "auto")
+        # Ask the user for the mode of operation for this auto step
+        is_interactive = confirm("Run with user confirmation for each tool execution?")
+        
+        if is_interactive:
+            self.console.print("Running a semi-automated step with user confirmation...", style="bold yellow")
+        else:
+            self.console.print("Running in fully automated mode...", style="bold yellow")
+        try:
+            # Pass the mode to the async function
+            asyncio.run(self.run_single_auto_step(interactive=is_interactive))
+        except Exception as e:
+            self.console.print(f"An error occurred during the automated step: {e}", style="bold red")
+            # In a real app, you'd use your logger here.
+            # logger.error(f"Auto step error: {traceback.format_exc()}")
+        response = "Automated step finished. Please review the results and choose your next action."
+        self.console.print(response, style="bold green")
+
+    elif request_option == "quit":
+        response = False
+        self.console.print("Thank you for using CTFGPT!", style="bold green")
+        self.log_conversation("ctfGPT", "Thank you for using CTFGPT!")
+
+    else:
+        response = "Please key in the correct options."
+        self.console.print(response, style="bold red")
+        self.log_conversation("ctfGPT", response)
+
+    return response
