+++ utils/ctf_gpt.py
@@ -1,8 +1,6 @@
-# an automated penetration testing parser empowered by GPT
 import json
 import os
 import sys
-import textwrap
 import time
 import traceback
 import asyncio 
@@ -14,50 +12,49 @@ from prompt_toolkit.shortcuts import confirm, prompt
 from rich.console import Console
 from rich.spinner import Spinner
 
-from config.chat_config import ChatGPTConfig
+from config.chat_config import ChatGPTConfig # Not directly used, but kept for context if needed later
 from prompts.prompt_class import CTFGPTPrompt
 from utils.APIs.module_import import dynamic_import
-# from utils.chatgpt import ChatGPT
-from utils.prompt_select import prompt_ask, prompt_select
-from utils.task_handler import (
-    local_task_entry,
-    localTaskCompleter,
-    main_task_entry,
-    mainTaskCompleter,
-)
 from utils.mcp_manager import MCPManager
+from utils.prompt_select import prompt_ask # Used by some handlers
+
+# Import handlers from the new cg_* files
+from .cg_constants import postfix_options, options_desc, prompt_continuation
+from .cg_session_handlers import log_conversation, save_session, _preload_session
+from .cg_llm_handlers import _feed_init_prompts, reasoning_handler, input_parsing_handler, test_generation_handler
+from .cg_ui_handlers import _select_mcp_server, local_input_handler, input_handler
+from .cg_auto_handlers import run_single_auto_step
+
 
 logger = loguru.logger
 
 
-def prompt_continuation(width, line_number, wrap_count):
-    """
-    The continuation: display line numbers and '->' before soft wraps.
-    Notice that we can return any kind of formatted text from here.
-    The prompt continuation doesn't have to be the same width as the prompt
-    which is displayed before the first line, but in this example we choose to
-    align them. The `width` input that we receive here represents the width of
-    the prompt.
-    """
-    if wrap_count > 0:
-        return " " * (width - 3) + "-> "
-    text = ("- %i - " % (line_number + 1)).rjust(width)
-    return HTML("<strong>%s</strong>") % text
+class CTFGPT:
+    # Constants moved to cg_constants.py, but keep reference for clarity if needed by other parts of the class
+    # or ensure they are passed/accessed correctly by the handler functions.
+    # For now, handler functions will access them via `self.postfix_options` etc. if they are set in __init__.
+    # To make them directly usable by imported functions if they weren't methods,
+    # they'd need to be imported in those files or passed as args.
+    # Since we are making them methods of CTFGPT by passing `self`, this is fine.
 
-class ctfGPT:
-    postfix_options = {
-        "tool": "The input content is from a security tool. You need to list down all the points that are interesting to you; you should summarize it as if you are reporting to a senior CTF player for further guidance.\n",
-        "user-comments": "The input content is from user comments.\n",
-        "web": "The input content is from web pages. You need to summarize the readable-contents, and list down all the points that can be interesting for the CTF.\n",
-        "default": "The user did not specify the input source. You need to summarize based on the contents.\n",
-    }
-
-    options_desc = {
-        "tool": " Paste the output of the CTF tool used",
-        "user-comments": "",
-        "web": " Paste the relevant content of a web page",
-        "default": " Write whatever you want, the tool will handle it",
-    }
+    # Assign imported handler functions as methods of the class
+    _feed_init_prompts = _feed_init_prompts
+    reasoning_handler = reasoning_handler
+    input_parsing_handler = input_parsing_handler
+    test_generation_handler = test_generation_handler
+    log_conversation = log_conversation
+    save_session = save_session
+    _preload_session = _preload_session
+    _select_mcp_server = _select_mcp_server
+    local_input_handler = local_input_handler
+    input_handler = input_handler
+    run_single_auto_step = run_single_auto_step
+    
+    # Constants that cg_llm_handlers and cg_ui_handlers might expect on self
+    # These were originally defined in the CTFGPT class body
+    postfix_options = postfix_options 
+    options_desc = options_desc
+    # prompt_continuation is also in cg_constants, used by prompt_ask if configured globally
 
     def __init__(
         self,
@@ -68,36 +65,42 @@ class ctfGPT:
         mcp_config_path="mcp_servers.json",
     ):
         self.log_dir = log_dir
+        # Ensure the log directory exists
+        os.makedirs(self.log_dir, exist_ok=True)
         logger.add(sink=os.path.join(log_dir, "ctfGPT.log"))
-        self.save_dir = "test_history"
-        self.task_log = (
-            {}
-        )  # the information that can be saved to continue in the next session
-        self.parsing_char_window = 16000  # the chunk size for parsing in # of chars
-        # TODO: link the parsing_char_window to the model used
-        # load the module
+        
+        self.save_dir = "test_history" # Ensure this directory is also handled (e.g., created in save_session)
+        os.makedirs(self.save_dir, exist_ok=True)
+
+        self.task_log = {}
+        self.parsing_char_window = 16000
+        
         reasoning_model_object = dynamic_import(
             reasoning_model, self.log_dir, use_langfuse_logging=use_langfuse_logging
         )
+        # Assuming generation model is same as reasoning, if not, it might need its own parameter
         generation_model_object = dynamic_import(
             reasoning_model, self.log_dir, use_langfuse_logging=use_langfuse_logging
         )
         parsing_model_object = dynamic_import(
             parsing_model, self.log_dir, use_langfuse_logging=use_langfuse_logging
         )
+        
         self.parsingAgent = parsing_model_object
         self.generationAgent = generation_model_object
         self.reasoningAgent = reasoning_model_object
-        self.prompts = CTFGPTPrompt
+        
+        self.prompts = CTFGPTPrompt # Class containing prompt strings
         self.console = Console()
-        self.spinner = Spinner("line", "Processing")
+        self.spinner = Spinner("line", "Processing") # Not explicitly used in provided snippets, but good to have
+        
         self.test_generation_session_id = None
         self.test_reasoning_session_id = None
         self.input_parsing_session_id = None
+        
         self.chat_count = 0
-        self.step_reasoning = (
-            None  # the response from the reasoning session for the current step
-        )
+        self.step_reasoning = None
+        
         self.history = {
             "user": [],
             "ctfGPT": [],
@@ -106,7 +109,12 @@ class ctfGPT:
             "generation": [],
             "mcp_tool_log": [],
             "exception": [],
-        }  # the history of the current conversation
+        }
+
+        self.mcp_manager = MCPManager(config_path=mcp_config_path)
+
+        # For cg_llm_handlers.py and others that might use prompt_ask directly or indirectly
+        self.prompt_ask = prompt_ask 
 
         # print the initialization message on the current implementation.
         self.console.print(
@@ -114,847 +122,114 @@ class ctfGPT:
             style="bold green",
         )
         self.console.print("The settings are: ")
-        self.console.print(
-            f" - parsing model: {parsing_model_object.name}", style="bold green"
-        )
-        self.console.print(
-            f" - reasoning model: {reasoning_model_object.name}", style="bold green"
-        )
+        self.console.print(f" - parsing model: {getattr(parsing_model_object, 'name', 'N/A')}", style="bold green")
+        self.console.print(f" - reasoning model: {getattr(reasoning_model_object, 'name', 'N/A')}", style="bold green")
         self.console.print(f" - log directory: {log_dir}", style="bold green")
-        self.mcp_manager = MCPManager(config_path=mcp_config_path)
+        self.console.print(f" - MCP config: {mcp_config_path}", style="bold green")
 
-    def log_conversation(self, source, text):
-        """
-        append the conversation into the history
 
-        Parameters:
-        ----------
-        source: str
-            the source of the conversation
-        text: str
-            the content of the conversation
-        """
-        # append the conversation into the history
-        timestamp = time.time()
-        if source not in self.history.keys():
-            # an exception
-            source = "exception"
-        self.history[source].append((timestamp, text))
-
-    def _feed_init_prompts(self):
-        # 1. User firstly provide basic information of the task
-        init_description = prompt_ask(
-            "Please describe the CTF challenge in one line, including the target IP, task type, etc.\n> ",
-            multiline=False,
-        )
-        self.log_conversation("user", init_description)
-        self.task_log["task description"] = init_description
-        # 2. Provide the information to the reasoning session for the task initialization.
-        # Note that this information is not parsed by the three-step process in reasoning.
-        # It is directly used to initialize the task.
-        prefixed_init_description = self.prompts.task_description + init_description
-        with self.console.status(
-            "[bold green] Constructing Initial CTF Task Tree..."
-        ) as status:
-            _reasoning_response = self.reasoningAgent.send_message(
-                prefixed_init_description, self.test_reasoning_session_id
-            )
-        # 3. Pass to generation session for more details.
-        # Note that the generation session is not used for the task initialization.
-        with self.console.status("[bold green] Generating Initial Task") as status:
-            _generation_response = self.generationAgent.send_message(
-                self.prompts.todo_to_command + _reasoning_response,
-                self.test_generation_session_id,
-            )
-        
-        self.step_reasoning = _reasoning_response 
-
-        # Display the initial generation result
-        response = _reasoning_response + "\n" + _generation_response
-        self.console.print("CTFGPT output: ", style="bold green")
-        self.console.print(response)
-        self.log_conversation("ctfGPT", "CTFGPT output:" + response)
-        
-    def _select_mcp_server(self) -> str:
-        """사용 가능한 MCP 서버 목록을 표시하고 사용자에게 선택을 요청합니다."""
-        available_servers = self.mcp_manager.get_available_servers()
-        
-        if not available_servers:
-            self.console.print("No available MCP servers found in the configuration file.", style="bold red")
-            return None
-
-        if len(available_servers) == 1:
-            server = available_servers[0]
-            self.console.print(f"Automatically selecting the only available MCP server: '{server}'", style="bold yellow")
-            return server
-            
-        # 사용자 선택 UI 생성
-        server_options = [
-            (str(i), server) for i, server in enumerate(available_servers)
-        ]
-        selected_idx_str = prompt_select(
-            title="Please select an MCP server to use:",
-            values=server_options
-        )
-        
-        if selected_idx_str is None:
-            return None
-            
-        return available_servers[int(selected_idx_str)]
-    
     def initialize(self, previous_session_ids=None):
         # initialize the backbone sessions and test the connection to chatGPT
         # define three sessions: testGenerationSession, testReasoningSession, and InputParsingSession
-        if previous_session_ids is not None:
-            self.test_generation_session_id = previous_session_ids.get(
-                "test_generation", None
-            )
-            self.test_reasoning_session_id = previous_session_ids.get("reasoning", None)
-            self.input_parsing_session_id = previous_session_ids.get("parsing", None)
-            # debug the three sessions
-            print("-----------------")
+        if previous_session_ids:
+            self.test_generation_session_id = previous_session_ids.get("test_generation")
+            self.test_reasoning_session_id = previous_session_ids.get("reasoning")
+            self.input_parsing_session_id = previous_session_ids.get("parsing")
             self.task_log = previous_session_ids.get("task_log", {})
-            self.console.print(f"Task log: {str(self.task_log)}", style="bold green")
-            print("You may use discussion function to remind yourself of the task.")
+            self.console.print(f"Loaded task log: {str(self.task_log)}", style="bold green")
+            self.console.print("You may use discussion function to remind yourself of the task.", style="bold yellow")
 
-            ## verify that all the sessions are not None
-            if (
-                self.test_generation_session_id is None
-                or self.test_reasoning_session_id is None
-                or self.input_parsing_session_id is None
-            ):
+            if not all([self.test_generation_session_id, self.test_reasoning_session_id, self.input_parsing_session_id]):
                 self.console.print(
-                    "[bold red] Error: the previous session ids are not valid. Loading new sessions"
+                    "[bold red]Error: Some previous session IDs are not valid. Initializing new sessions.[/bold red]"
                 )
-                self.initialize()
-
-        else:
-            with self.console.status(
-                "[bold green] Initialize ChatGPT Sessions..."
-            ) as status:
-                try:
-                    (
-                        text_0,
-                        self.test_generation_session_id,
-                    ) = self.generationAgent.send_new_message(
-                        self.prompts.generation_session_init,
-                    )
-                    (
-                        text_1,
-                        self.test_reasoning_session_id,
-                    ) = self.reasoningAgent.send_new_message(
-                        self.prompts.reasoning_session_init
-                    )
-                    (
-                        text_2,
-                        self.input_parsing_session_id,
-                    ) = self.parsingAgent.send_new_message(
-                        self.prompts.input_parsing_init
-                    )
-                except Exception as e:
-                    logger.error(e)
-            self.console.print("- ChatGPT Sessions Initialized.", style="bold green")
-            self._feed_init_prompts()
-
-    def reasoning_handler(self, text) -> str:
-        # summarize the contents if necessary.
-        if len(text) > self.parsing_char_window:
-            text = self.input_parsing_handler(text)
-        """
-        # pass the information to reasoning_handler and obtain the results
-        response = self.reasoningAgent.send_message(
-            self.prompts.process_results + text, self.test_reasoning_session_id
-        )
-        # log the conversation
-        """
-        # CTFGPT Reasoning Logic
-        ## 1. Given the information, update the CTT
-        _updated_ptt_response = self.reasoningAgent.send_message(
-            self.prompts.process_results + text, self.test_reasoning_session_id
-        )
-        ## 2. Validate if the CTT is correct
-        # TODO
-        ## 3. If the CTT is correct, select all the to-dos
-        _task_selection_response = self.reasoningAgent.send_message(
-            self.prompts.process_results_task_selection, self.test_reasoning_session_id
-        )
-        # get the complete output:
-        response = _updated_ptt_response + _task_selection_response
-
-        self.log_conversation("reasoning", response)
-        return response
-
-    def input_parsing_handler(self, text, source=None) -> str:
-        prefix = "Please summarize the following input. "
-        # do some engineering trick here. Add postfix to the input to make it more understandable by LLMs.
-        if source is not None and source in self.postfix_options.keys():
-            prefix += self.postfix_options[source]
-        # The default token-size limit is 4096 (web UI even shorter). 1 token ~= 4 chars in English
-        # Use textwrap to split inputs. Limit to 2000 token (8000 chars) for each input
-        # (1) replace all the newlines with spaces
-        text = text.replace("\r", " ").replace("\n", " ")
-        # (2) wrap the text
-        wrapped_text = textwrap.fill(text, 8000)
-        wrapped_inputs = wrapped_text.split("\n")
-        # (3) send the inputs to chatGPT input_parsing_session and obtain the results
-        summarized_content = ""
-        for wrapped_input in wrapped_inputs:
-            word_limit = f"Please ensure that the input is less than {8000 / len(wrapped_inputs)} words.\n"
-            summarized_content += self.parsingAgent.send_message(
-                prefix + word_limit + wrapped_input, self.input_parsing_session_id
-            )
-        # log the conversation
-        self.log_conversation("input_parsing", summarized_content)
-        return summarized_content
-
-    def test_generation_handler(self, text):
-        # send the contents to chatGPT test_generation_session and obtain the results
-        response = self.generationAgent.send_message(
-            text, self.test_generation_session_id
-        )
-        # log the conversation
-        self.log_conversation("generation", response)
-        return response
-
-    def local_input_handler(self) -> str:
-        """
-        Request for user's input to handle the local task
-        """
-        local_task_response = ""
-        self.chat_count += 1
-        local_request_option = local_task_entry()
-        self.log_conversation("user", local_request_option)
-
-        if local_request_option == "help":
-            print(localTaskCompleter().task_details)
-
-        elif local_request_option == "discuss":
-            ## (1) Request for user multi-line input
-            self.console.print(
-                "Please share your findings and questions with CTFGPT."
-            )
-            self.log_conversation(
-                "ctfGPT",
-                "Please share your findings and questions with CTFGPT. (End with <shift + right-arrow>)",
-            )
-            user_input = prompt_ask("Your input: ", multiline=True)
-            self.log_conversation("user", user_input)
-            ## (2) pass the information to the reasoning session.
-            with self.console.status("[bold green] CTFGPT Thinking...") as status:
-                local_task_response = self.test_generation_handler(
-                    self.prompts.local_task_prefix + user_input
-                )
-            ## (3) print the results
-            self.console.print("CTFGPT:\n", style="bold green")
-            self.console.print(local_task_response + "\n", style="yellow")
-            self.log_conversation("ctfGPT", local_task_response)
-
-        elif local_request_option == "brainstorm":
-            ## (1) Request for user multi-line input
-            self.console.print(
-                "Please share your concerns and questions with CTFGPT."
-            )
-            self.log_conversation(
-                "ctfGPT",
-                "Please share your concerns and questions with CTFGPT. End with <shift + right-arrow>)",
-            )
-            user_input = prompt_ask("Your input: ", multiline=True)
-            self.log_conversation("user", user_input)
-            ## (2) pass the information to the reasoning session.
-            with self.console.status("[bold green] CTFGPT Thinking...") as status:
-                local_task_response = self.test_generation_handler(
-                    self.prompts.local_task_brainstorm + user_input
-                )
-            ## (3) print the results
-            self.console.print("CTFGPT:\n", style="bold green")
-            self.console.print(local_task_response + "\n", style="yellow")
-            self.log_conversation("ctfGPT", local_task_response)
-
-        elif local_request_option == "continue":
-            self.console.print("Exit the local task and continue the CTF.")
-            self.log_conversation(
-                "ctfGPT", "Exit the local task and continue the CTF."
-            )
-            local_task_response = "continue"
-
-        return local_task_response
-
-    def input_handler(self) -> str:
-        """
-        Request for user's input to:
-            (1) input test results,
-            (2) ask for todos,
-            (3) input other information (discuss),
-            (4) end.
-        The design details are based on CTFGPT_design.md
-
-        Return
-        -----
-        response: str
-            The response from the chatGPT model.
-        """
-        self.chat_count += 1
-
-        request_option = main_task_entry()
-        self.log_conversation("user", request_option)
-
-        if request_option == "help":
-            print(mainTaskCompleter().task_details)
-
-        if request_option == "next":
-            ## (1) pass the information to input_parsing session.
-            ## Give an option list for user to choose from
-            options = list(self.postfix_options.keys())
-            opt_desc = list(self.options_desc.values())
-
-            value_list = [
-                (
-                    i,
-                    HTML(
-                        f'<style fg="cyan">{options[i]}</style><style fg="LightSeaGreen">{opt_desc[i]}</style>'
-                    ),
-                )
-                for i in range(len(options))
-            ]
-            source = prompt_select(
-                title="Please choose the source of the information.", values=value_list
-            )
-            self.console.print(
-                "Your input: (End with <shift + right-arrow>)", style="bold green"
-            )
-            user_input = prompt_ask("> ", multiline=True)
-            self.log_conversation(
-                "user", f"Source: {options[int(source)]}" + "\n" + user_input
-            )
-            with self.console.status("[bold green] CTFGPT Thinking...") as status:
-                parsed_input = self.input_parsing_handler(
-                    user_input, source=options[int(source)]
-                )
-                ## (2) pass the summarized information to the reasoning session.
-                reasoning_response = self.reasoning_handler(parsed_input)
-                self.step_reasoning = reasoning_response
-
-            ## (3) print the results
-            self.console.print(
-                "Based on the analysis, the following tasks are recommended:",
-                style="bold green",
-            )
-            self.console.print(reasoning_response + "\n")
-            self.log_conversation(
-                "ctfGPT",
-                "Based on the analysis, the following tasks are recommended:"
-                + reasoning_response,
-            )
-            response = reasoning_response
-
-        elif request_option == "more":
-            self.log_conversation("user", "more")
-            ## (1) check if reasoning session is initialized
-            if not self.step_reasoning:
-                self.console.print(
-                    "You have not initialized the task yet. Please perform the basic recon following `next` option.",
-                    style="bold red",
-                )
-                response = "You have not initialized the task yet. Please perform the basic recon following `next` option."
-                self.log_conversation("ctfGPT", response)
-                return response
-            ## (2) start local task generation.
-            ### (2.1) ask the reasoning session to analyze the current situation, and explain the task
-            self.console.print(
-                "CTFGPT will generate more step details, and enter the sub-task generation mode. (Pressing Enter to continue)",
-                style="bold green",
-            )
-            self.log_conversation(
-                "ctfGPT",
-                "CTFGPT will generate more step details, and enter the sub-task generation mode.",
-            )
-            input()
-
-            ### (2.2) pass the sub-tasks to the test generation session
-            with self.console.status("[bold green] CTFGPT Thinking...") as status:
-                generation_response = self.test_generation_handler(
-                    self.step_reasoning
-                )
-                _local_init_response = self.test_generation_handler(
-                    self.prompts.local_task_init
-                )
-
-            self.console.print(
-                "Below are the further details.",
-                style="bold green",
-            )
-            self.console.print(generation_response + "\n")
-            response = generation_response
-            self.log_conversation("ctfGPT", response)
-
-            ### (2.3) local task handler
-            while True:
-                local_task_response = self.local_input_handler()
-                if local_task_response == "continue":
-                    # break the local task handler
-                    break
-
-        elif request_option == "todo":
-            ## log that user is asking for todo list
-            self.log_conversation("user", "todo")
-            ## (1) ask the reasoning session to analyze the current situation, and list the top sub-tasks
-            with self.console.status("[bold green] CTFGPT Thinking...") as status:
-                reasoning_response = self.reasoning_handler(self.prompts.ask_todo)
-                self.step_reasoning = reasoning_response 
-                ## (2) pass the sub-tasks to the test_generation session.
-                message = self.prompts.todo_to_command + "\n" + reasoning_response
-                generation_response = self.test_generation_handler(message)
-                ## (3) print the results
-            self.console.print(
-                "Based on the analysis, the following tasks are recommended:",
-                style="bold green",
-            )
-            self.console.print(reasoning_response + "\n")
-            self.console.print(
-                "You can follow the instructions below to complete the tasks.",
-                style="bold green",
-            )
-            self.console.print(generation_response + "\n")
-            response = reasoning_response
-            self.log_conversation(
-                "ctfGPT",
-                (
-                    (
-                        (
-                            (
-                                "Based on the analysis, the following tasks are recommended:"
-                                + response
-                            )
-                            + "\n"
-                        )
-                        + "You can follow the instructions below to complete the tasks."
-                    )
-                    + generation_response
-                ),
-            )
-        elif request_option == "discuss":
-            ## (1) Request for user multi-line input
-            self.console.print(
-                "Please share your thoughts/questions with CTFGPT. (End with <shift + right-arrow>) "
-            )
-            self.log_conversation(
-                "ctfGPT", "Please share your thoughts/questions with CTFGPT."
-            )
-            user_input = prompt_ask("Your input: ", multiline=True)
-            self.log_conversation("user", user_input)
-            ## (2) pass the information to the reasoning session.
-            with self.console.status("[bold green] CTFGPT Thinking...") as status:
-                response = self.reasoning_handler(self.prompts.discussion + user_input)
-            ## (3) print the results
-            self.console.print("CTFGPT:\n", style="bold green")
-            self.console.print(response + "\n", style="yellow")
-            self.log_conversation("ctfGPT", response)
-
-        elif request_option == "auto":
-            self.log_conversation("user", "auto")
-
-            # Ask the user for the mode of operation for this auto step
-            is_interactive = confirm("Run with user confirmation for each tool execution?")
-            
-            if is_interactive:
-                self.console.print("Running a semi-automated step with user confirmation...", style="bold yellow")
+                # Fall through to initialize new sessions
+                previous_session_ids = None # Force re-initialization
             else:
-                self.console.print("Running in fully automated mode...", style="bold yellow")
+                self.console.print("- Loaded sessions from previous state.", style="bold green")
+                return # Successfully loaded
 
+        # Initialize new sessions if not loaded from previous
+        with self.console.status("[bold green]Initializing LLM Sessions...[/bold green]") as status:
             try:
-                # Pass the mode to the async function
-                asyncio.run(self.run_single_auto_step(interactive=is_interactive))
-
-            except Exception as e:
-                self.console.print(f"An error occurred during the automated step: {e}", style="bold red")
-                logger.error(f"Auto step error: {traceback.format_exc()}")
-
-            response = "Automated step finished. Please review the results and choose your next action."
-            self.console.print(response, style="bold green")
-
-        elif request_option == "quit":
-            response = False
-            self.console.print("Thank you for using CTFGPT!", style="bold green")
-            self.log_conversation("ctfGPT", "Thank you for using CTFGPT!")
-
-        else:
-            self.console.print("Please key in the correct options.", style="bold red")
-            self.log_conversation("ctfGPT", "Please key in the correct options.")
-            response = "Please key in the correct options."
-        return response
-
-    def save_session(self):
-        """
-        Save the current session for next round of usage.
-        The test information is saved in the directory `./test_history`
-        """
-        self.console.print(
-            "Before you quit, you may want to save the current session.",
-            style="bold green",
-        )
-        # 1. Require a save name from the user. If not, use the current time as the save name.
-        save_name = prompt_ask(
-            "Please enter the name of the current session. (Default with current timestamp)\n> ",
-            multiline=False,
-        )
-        if save_name == "":
-            save_name = str(time.time())
-        # 2. Save the current session
-        with open(
-            os.path.join(
-                os.path.realpath(os.path.dirname(__file__)),
-                os.pardir,
-                os.pardir,
-                self.save_dir,
-                save_name,
-            ),
-            "w",
-        ) as f:
-            # store the three ids and task_log
-            session_ids = {
-                "reasoning": self.test_reasoning_session_id,
-                "test_generation": self.test_generation_session_id,
-                "parsing": self.input_parsing_session_id,
-                "task_log": self.task_log,
-            }
-            json.dump(session_ids, f)
-        self.console.print(
-            f"The current session is saved as {save_name}", style="bold green"
-        )
-        return
-
-    def _preload_session(self) -> dict:
-        """
-        Preload the session from the save directory.
-
-        Returns:
-            dict: the session ids for the three sessions.
-            None if no previous session is found.
-        """
-        if continue_from_previous := confirm(
-            "Do you want to continue from previous session?"
-        ):
-            # load the filenames from the save directory
-            filenames = os.listdir(
-                os.path.join(
-                    os.path.realpath(os.path.dirname(__file__)),
-                    os.pardir,
-                    os.pardir,
-                    self.save_dir,
+                # generation session
+                _, self.test_generation_session_id = self.generationAgent.send_new_message(
+                    self.prompts.generation_session_init,
                 )
-            )
-            if len(filenames) == 0:
-                print("No previous session found. Please start a new session.")
-                return None
-            else:  # print all the files
-                print("Please select the previous session by its index (integer):")
-                for i, filename in enumerate(filenames):
-                    print(f"{str(i)}. {filename}")
-                # ask for the user input
-                try:
-                    previous_testing_name = filenames[
-                        int(input("Please key in your option (integer): "))
-                    ]
-                    print(f"You selected: {previous_testing_name}")
-                except ValueError as e:
-                    print("You input an invalid option. Will start a new session.")
-                    return None
-
-        elif continue_from_previous is False:
-            return None
-        else:
-            print("You input an invalid option. Will start a new session.")
-            return None
-        # 2. load the previous session information
-        if previous_testing_name is not None:
-            # try to load the file content with json
-            try:
-                with open(
-                    os.path.join(
-                        os.path.realpath(os.path.dirname(__file__)),
-                        os.pardir,
-                        os.pardir,
-                        self.save_dir,
-                        previous_testing_name,
-                    ),
-                    "r",
-                ) as f:
-                    return json.load(f)
-            except Exception as e:
-                print(
-                    "Error when loading the previous session. The file name is not correct"
+                # reasoning session
+                _, self.test_reasoning_session_id = self.reasoningAgent.send_new_message(
+                    self.prompts.reasoning_session_init
                 )
-                print(e)
-                previous_testing_name = None
-                return None
+                # parsing session
+                _, self.input_parsing_session_id = self.parsingAgent.send_new_message(
+                    self.prompts.input_parsing_init
+                )
+                self.console.print("- LLM Sessions Initialized.", style="bold green")
+            except Exception as e:
+                logger.error(f"Failed to initialize LLM sessions: {e}")
+                self.console.print(f"[bold red]Error initializing LLM sessions: {e}[/bold red]")
+                # Depending on severity, might want to exit or let user retry
+                raise # Re-raise for now to make failure explicit
 
-    async def run_single_auto_step(self, interactive: bool):
-        """
-        Executes a single automated step (tool execution) based on the LLM's suggestion.
-        The behavior differs based on the 'interactive' flag.
-        :param interactive: If True, the user selects the tool and confirms arguments.
-                            If False, the LLM selects the server, tool, and arguments.
-        """
-        self.console.print("\n--- Preparing for automated step ---", style="bold magenta")
-        
-        if not self.step_reasoning:
-            self.console.print("No task has been suggested yet. Please run 'next' or 'todo' first.", style="bold red")
-            return
-        
-        task_to_execute = self.step_reasoning
+        # After initializing new sessions, feed initial prompts if it's a completely new start
+        if not previous_session_ids: # Only if we didn't attempt to load a session or loading failed
+             self._feed_init_prompts()
 
-        # 1. Fetch all available tools from all servers first.
-        with self.console.status("[bold green]Fetching tools from all available MCP servers..."):
-            all_tools_by_server = {}
-            server_names = self.mcp_manager.get_available_servers()
-            for server_name in server_names:
-                client = self.mcp_manager.create_client(server_name)
-                if not client: continue
-                try:
-                    async with client:
-                        if client.is_connected():
-                            tools = await client.list_tools()
-                            all_tools_by_server[server_name] = [
-                                {"name": tool.name, "description": tool.description, "inputSchema": tool.inputSchema} for tool in tools
-                            ]
-                except Exception as e:
-                    self.console.print(f"Warning: Could not fetch tools from '{server_name}': {e}", style="yellow")
-        
-        if not all_tools_by_server:
-            self.console.print("Could not fetch any tools from any MCP server. Aborting.", style="bold red")
-            return
-
-        action = {}
-        # 2. Branch logic based on interactive mode
-        if interactive:
-            # --- Interactive Mode ---
-            # 2a. User selects the tool
-            self.console.print("Please select a tool to use for the task:", style="bold green")
-            tool_options = []
-            for server, tools in all_tools_by_server.items():
-                for tool in tools:
-                    # Value format: "server_name.tool_name"
-                    tool_id = f"{server}.{tool['name']}"
-                    display_text = HTML(f'<style fg="cyan">{server}</style>.<style fg="yellow">{tool["name"]}</style>: {tool["description"]}')
-                    tool_options.append((tool_id, display_text))
-
-            selected_tool_id = await asyncio.to_thread(prompt_select, title="Available Tools:", values=tool_options)
-            
-            if not selected_tool_id:
-                self.console.print("No tool selected. Aborting auto step.", style="yellow")
-                return
-            
-            selected_server_name, selected_tool_name = selected_tool_id.split('.', 1)
-            selected_tool_info = next(t for t in all_tools_by_server[selected_server_name] if t['name'] == selected_tool_name)
-
-            # 2b. LLM generates arguments for the selected tool
-            prompt_for_args = f"""
-            You are an expert at generating command arguments.
-            The user wants to perform the following task:
-            ---
-            {task_to_execute}
-            ---
-            The user has chosen to use the '{selected_tool_name}' tool from the '{selected_server_name}' server.
-            Here is the tool's information:
-            ```json
-            {json.dumps(selected_tool_info, indent=2)}
-            ```
-            Based on the task and the tool's input schema, generate the appropriate JSON object for the 'arguments' key.
-            Your output MUST be only a single JSON object containing the arguments.
-
-            Example output for nmap_scan:
-            {{"targets": "localhost", "options": "-p-"}}
-            """
-            if False:
-                # --- DEBUG: Print prompt for LLM ---
-                self.console.print("\n--- DEBUG: Prompt to LLM ---", style="yellow")
-                self.console.print(prompt_for_args)
-                self.console.print("--------------------------\n", style="yellow")
-
-            with self.console.status(f"[bold green]LLM is generating arguments for {selected_tool_name}..."):
-                args_json_str = self.reasoningAgent.send_message(prompt_for_args, self.test_reasoning_session_id)
-
-            if False:
-                # --- DEBUG: Print raw LLM response ---
-                self.console.print("\n--- DEBUG: Raw LLM Response ---", style="yellow")
-                self.console.print(args_json_str)
-                self.console.print("-----------------------------\n", style="yellow")
-            
-            try:
-                json_match = re.search(r'\{.*\}', args_json_str, re.DOTALL)
-                arguments = json.loads(json_match.group())
-                action = {
-                    "server_name": selected_server_name,
-                    "tool_name": selected_tool_name,
-                    "arguments": arguments
-                }
-            except (json.JSONDecodeError, AttributeError) as e:
-                self.console.print(f"Failed to parse arguments from LLM: {e}", style="bold red")
-                self.console.print("LLM Response:", args_json_str, style="red")
-                return
-
-        else:
-            # --- Fully Automated Mode ---
-            # 2c. LLM selects server, tool, and arguments in one go
-            all_tools_json_str = json.dumps(all_tools_by_server, indent=2)
-            prompt_for_full_action = f"""
-            You are an expert CTF assistant responsible for selecting the best tool for a given task from multiple available tool servers.
-
-            Recommended Task:
-            ---
-            {task_to_execute}
-            ---
-
-            Available Tool Servers and their Tools:
-            ```json
-            {all_tools_json_str}
-            ```
-
-            Based on the "Recommended Task", analyze the tools available on each server. Select the single most appropriate server and tool to execute the task.
-            Your output MUST be a single JSON object with "server_name", "tool_name", and "arguments" keys.
-
-            IMPORTANT: If no tool on any server is suitable, return a JSON object with "server_name" set to "none".
-
-            Example for a port scan task:
-            {{"server_name": "ctf-tools", "tool_name": "nmap_scan", "arguments": {{"targets": "localhost", "options": "-p-"}}}}
-
-            Example for when no tool is suitable:
-            {{"server_name": "none", "tool_name": "none", "arguments": {{"reason": "The task requires a specific vulnerability scanner that is not available."}}}}
-            """
-            if False:
-                # --- DEBUG: Print prompt for LLM ---
-                self.console.print("\n--- DEBUG: Prompt to LLM ---", style="yellow")
-                self.console.print(prompt_for_full_action)
-                self.console.print("--------------------------\n", style="yellow")
-
-            with self.console.status("[bold green]LLM is selecting the best server, tool, and arguments..."):
-                action_json_str = self.reasoningAgent.send_message(prompt_for_full_action, self.test_reasoning_session_id)
-
-            if False:
-                # --- DEBUG: Print raw LLM response ---
-                self.console.print("\n--- DEBUG: Raw LLM Response ---", style="yellow")
-                self.console.print(action_json_str)
-                self.console.print("-----------------------------\n", style="yellow")
-
-            try:
-                json_match = re.search(r'\{.*\}', action_json_str, re.DOTALL)
-                if not json_match: 
-                    raise ValueError("No JSON object found in the response.")
-                action = json.loads(json_match.group())
-            except (json.JSONDecodeError, AttributeError) as e:
-                self.console.print(f"Failed to parse action from LLM: {e}", style="bold red")
-                self.console.print("LLM Response:", action_json_str, style="red")
-                return
-
-        # 3. Execution (common for both modes)
-        server_name = action.get("server_name")
-        tool_name = action.get("tool_name")
-        arguments = action.get("arguments")
-
-        if not server_name or server_name == 'none':
-            self.console.print("LLM determined no suitable tool is available or chose not to act.", style="yellow")
-            return
-
-        self.console.print(f"Suggested Action: Run tool '{tool_name}' on server '{server_name}' with arguments: {arguments}", style="cyan")
-
-        # Final confirmation step (always happens in interactive mode, skipped in full-auto)
-        execute = True
-        if interactive:
-            execute = await asyncio.to_thread(
-                confirm, f"Proceed with execution?"
-            )
-        
-        if not execute:
-            self.console.print("Execution cancelled by user.", style="yellow")
-            return
-
-        # 4. Connect to the chosen server and run the tool
-        client = self.mcp_manager.create_client(server_name)
-        if not client:
-            self.console.print(f"Failed to create client for the selected server '{server_name}'.", style="bold red")
-            return
-        
-        async with client:
-            if not client.is_connected():
-                self.console.print(f"MCP Client could not connect to '{server_name}'. Check server script and permissions.", style="bold red")
-                return
-            
-            with self.console.status(f"[bold green]Executing '{tool_name}' on '{server_name}'..."):
-                tool_results_list = await client.call_tool(tool_name, arguments)
-        
-        if not tool_results_list:
-            self.console.print(f"Tool '{tool_name}' executed but returned no result.", style="yellow")
-            return
-        
-        tool_output_text = "\n".join([result.text for result in tool_results_list])
-        
-        # 로깅
-        mcp_log_entry = {
-            "server": server_name, 
-            "tool": tool_name,
-            "arguments": arguments,
-            "result": tool_output_text
-        }
-        self.log_conversation("mcp_tool_log", json.dumps(mcp_log_entry, indent=2))
-        
-        self.console.print(f"--- Tool '{tool_name}' Output ---", style="green")
-        self.console.print(tool_output_text)
-        self.console.print("--- End of Output ---", style="green")
-        
-        # 5. Analyze results and update CTT
-        with self.console.status("[bold green]Analyzing results and updating PTT..."):
-            self.step_reasoning = self.reasoning_handler(tool_output_text)
-            self.console.print("\nCTFGPT's analysis and next recommended steps:", style="bold green")
-            self.console.print(self.step_reasoning)
-
-        self.console.print("CTT updated.", style="green")
 
     def main(self):
         """
         The main function of ctfGPT. The design is based on CTFGPT_design.md
         """
         # 0. initialize the backbone sessions and test the connection to chatGPT
-        loaded_ids = self._preload_session()
-        self.initialize(previous_session_ids=loaded_ids)
+        loaded_ids = self._preload_session() # This is now a method from cg_session_handlers
+        try:
+            self.initialize(previous_session_ids=loaded_ids)
+        except Exception as e:
+            self.console.print(f"Critical error during initialization: {e}", style="bold red")
+            logger.error(f"Initialization failed: {traceback.format_exc()}")
+            return
 
-        # enter the main loop.
         while True:
             try:
-                result = self.input_handler()
-                self.console.print(
-                    "-----------------------------------------", style="bold white"
-                )
-                if not result:  # end the session
+                # input_handler is now a method from cg_ui_handlers
+                result = self.input_handler() 
+                self.console.print("-----------------------------------------", style="bold white")
+                if result is False:  # Explicitly check for False to quit
                     break
-            except Exception as e:  # catch all general exception.
-                # log the exception
-                self.log_conversation("exception", str(e))
-                # print the exception
+            except Exception as e:
+                self.log_conversation("exception", traceback.format_exc()) # Log full traceback
                 self.console.print(f"Exception: {str(e)}", style="bold red")
-                # add a more detailed debugging
                 exc_type, exc_obj, exc_tb = sys.exc_info()
                 fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
                 self.console.print(
-                    "Exception details are below. You may submit an issue on github and paste the error trace",
-                    style="bold green",
+                    f"Error in {fname}, line {exc_tb.tb_lineno}. Traceback logged.",
+                    style="bold red",
                 )
-                # self.console.print(exc_type, fname, exc_tb.tb_lineno)
-                print(traceback.format_exc())
-                # safely quit the session
-                break
+                self.console.print("Submitting an issue on GitHub with the error trace from logs might help.", style="bold yellow")
+                # Consider if we should break or allow continuing, for now, break on unhandled exception in loop
+                break 
         
-        # log the session. Save self.history into a txt file based on timestamp
+        # Save history and session details
         timestamp = time.time()
-        log_name = f"ctfGPT_log_{str(timestamp)}.txt"
-        # save it in the logs folder
+        # Use self.log_dir for consistency
+        log_name = f"ctfGPT_log_{str(timestamp)}.json" 
         log_path = os.path.join(self.log_dir, log_name)
-        with open(log_path, "w") as f:
-            json.dump(self.history, f)
-        self.save_session()
+        try:
+            with open(log_path, "w") as f:
+                json.dump(self.history, f, indent=4)
+            self.console.print(f"Chat history saved to {log_path}", style="bold green")
+        except Exception as e:
+            self.console.print(f"Failed to save chat history: {e}", style="bold red")
+            logger.error(f"Failed to save chat history: {traceback.format_exc()}")
 
-        self.save_session()
-
+        self.save_session() # This is now a method from cg_session_handlers
 
 if __name__ == "__main__":
-    ctfGPT = ctfGPT()
-    ctfGPT.main()
+    # Example of how to run if this file itself is executed
+    # Typically, main() in main.py would handle this.
+    try:
+        ctf_gpt_instance = CTFGPT()
+        ctf_gpt_instance.main()
+    except Exception as e:
+        logger.error(f"ctfGPT failed to run: {traceback.format_exc()}")
+        print(f"An critical error occurred: {e}")
